pyproject.toml
[tool.poetry]
name = "adimen-test"
version = "0.1.0"
description = ""
authors = ["sergpsu <sergpsu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.0"
sqlmodel = "^0.0.16"
aioboto3 = "^12.3.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.3"}
aiosqlite = "^0.19.0"
pytest = "^8.0.1"
httpx = "^0.26.0"
anyio = "^4.3.0"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.13.1"
fastapi-users-db-sqlalchemy = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
